Extract avatar partial components to separate files
