+++ app/src/userpages/components/Avatar/AvatarUpload/avatarUpload.pcss
@@ -1,0 +1,10 @@
+.upload {
+  color: #323232;
+  font-size: 14px;
+  letter-spacing: 0;
+  line-height: 20px;
+}
+
+.uploadHelpText {
+  margin-top: 1rem;
+}
