+++ app/src/userpages/components/Avatar/NameAndEmail/nameAndEmail.pcss
@@ -1,0 +1,24 @@
+.content {
+  flex-grow: 1;
+}
+
+.name {
+  color: #323232;
+  font-size: 28px;
+  font-family: 'IBM Plex Sans', sans-serif;
+  letter-spacing: 0;
+  line-height: 32px;
+}
+
+.email {
+  display: inline-block;
+  color: #525252;
+  font-size: 12px;
+  font-family: 'IBM Plex Sans', sans-serif;
+  letter-spacing: 0.75px;
+  background: #EFEFEF;
+  border-radius: 2px;
+  padding: 6px;
+  line-height: 12px;
+  margin-top: 0.5rem;
+}
