+++ app/src/userpages/components/Avatar/avatar.pcss
@@ -8,39 +8,3 @@
   border-radius: 50%;
   margin-right: 2rem;
 }
-
-.content {
-  flex-grow: 1;
-}
-
-.name {
-  color: #323232;
-  font-size: 28px;
-  font-family: 'IBM Plex Sans', sans-serif;
-  letter-spacing: 0;
-  line-height: 32px;
-}
-
-.email {
-  display: inline-block;
-  color: #525252;
-  font-size: 12px;
-  font-family: 'IBM Plex Sans', sans-serif;
-  letter-spacing: 0.75px;
-  background: #EFEFEF;
-  border-radius: 2px;
-  padding: 6px;
-  line-height: 12px;
-  margin-top: 0.5rem;
-}
-
-.upload {
-  color: #323232;
-  font-size: 14px;
-  letter-spacing: 0;
-  line-height: 20px;
-}
-
-.uploadHelpText {
-  margin-top: 1rem;
-}
